Renaming list of MDX components in a new file
