+++ src/components/Mdx/Custom.tsx
@@ -1,8 +1,8 @@
 import React, { ReactNode, useContext } from 'react';
 import { PageProps } from 'gatsby';
-import { ThemeContext } from './Layout';
+import { ThemeContext } from '../Layout';
 import classNames from 'classnames';
-import Code from './Mdx/Code';
+import Code from './Code';
 import { getImage } from 'gatsby-plugin-image';
 
 interface Props {
@@ -86,6 +86,7 @@ export const MyPre = (props: { children: { props: Record<string, unknown> }; cla
 export const MyImage = (props: Record<string, unknown>) => <img className="shadow-lg rounded" alt="" {...props} />;
 
 export const MyPostLink = (props) => (
+  // eslint-disable-next-line jsx-a11y/anchor-has-content
   <a className={classNames('underline underline-offset-4')} {...props} target="_blank" />
 );
